Conflicts:
	src/codename/network/src/se/rhel/client/ClientPacketHandler.java
	src/codename/network/src/se/rhel/packet/RequestInitialStatePacket.java
	src/codename/network/src/se/rhel/server/Server.java
	src/codename/network/src/se/rhel/server/TcpConnection.java

+++ src/codename/network/src/se/rhel/packet/Packet.java
@@ -15,7 +15,7 @@ public abstract class Packet {
         mPacketId = (byte) packetId;
         mData = new byte[packetSize];
         mBuffer = ByteBuffer.wrap(mData);
-        mBuffer.put((byte)mPacketId);
+        mBuffer.put(mPacketId);
     }
 
     public byte getPacketId() {
